backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
pyyaml
python-dotenv
google-generativeai
PyPDF2
python-docx
flasgger
